+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+- New `Identifier::idIsEmpty()` static method for checking that an `id` value is not empty. This check ensures that the
+  trimmed string is empty and that the string is not `"0"`, as zero could be used for a resource id. This new static
+  method has been added to ensure the logic of determining if an id value is empty is in one place.
+
+### Fixed
+
+- The package now accepts `"0"` as a valid client generated id. Previously it was rejected with a message that the id
+  could not be empty.
+- The package now accepts a document for a resource with an expected id of `"0"`. Previously the package did not 
+  recognise zero as an expected id.
+
 ## [1.1.0] - 2022-02-09
 
 ### Added